+++ app/javascript/vue/components/radials/annotator/components/citations/edit.vue
@@ -1,6 +1,11 @@
 <template>
   <div class="field">
-    <span v-html="editCitation.object_tag"/>
+    <div class="horizontal-left-content">
+      <span v-html="editCitation.object_tag"/>
+      <soft-validation
+        class="margin-small-left"
+        :global-id="editCitation.global_id"/>
+    </div>
     <div class="flex-separate separate-bottom separate-top field">
       <label class="inline middle">
         <input
@@ -29,39 +34,43 @@
 </template>
 <script>
 
-  export default {
-    props: {
-      citation: {
-        type: Object,
-        required: true
-      },
-      globalId: {
-        type: String,
-        required: true
-      }
+import SoftValidation from 'components/soft_validations/objectValidation.vue'
+export default {
+  components: {
+    SoftValidation
+  },
+  props: {
+    citation: {
+      type: Object,
+      required: true
     },
-    computed: {
-      validateFields() {
-        return this.citation.source_id
-      }
+    globalId: {
+      type: String,
+      required: true
+    }
+  },
+  computed: {
+    validateFields() {
+      return this.citation.source_id
+    }
+  },
+  data() {
+    return {
+      editCitation: this.citation,
+    }
+  },
+  watch: {
+    citation(newVal) {
+      this.citation = newVal
+    }
+  },
+  methods: {
+    sendCitation() {
+      this.$emit('update', this.editCitation)
     },
-    data() {
-      return {
-        editCitation: this.citation,
-      }
-    },
-    watch: {
-      citation(newVal) {
-        this.citation = newVal
-      }
-    },
-    methods: {
-      sendCitation() {
-        this.$emit('update', this.editCitation)
-      },
-      newCitation() {
-        this.$emit('new', true)
-      }
+    newCitation() {
+      this.$emit('new', true)
     }
   }
+}
 </script>